Connection 동기화 관련 코드 개선.
- BlinkDevice의 GroupID 필드의 타입이 변경되었으므로, 
	안드로이드 로컬 디렉토리 ..Blink/system/dev를 지우고 사용해야함.
	
- Identity 동기화 > NetworkMap 동기화 > BlinkAppInfo 동기화 까지 순차적으로 동작함.
	(MessageProcessor 참고)

- 현재 Center가 연결되어 있는 디바이스 W1에 W2가 연결될 경우 (M-W1-W2), W2는 동기화가 불가능함.
	이 부분에 대한 동기화 로직이 필요할 듯..

+++ BlinkLibrary/src/kr/poturns/blink/internal/DeviceAnalyzer.java
@@ -279,7 +279,7 @@ public class DeviceAnalyzer {
 	 * @param device2
 	 * @return
 	 */
-	int compareForIdentity(BlinkDevice device1, BlinkDevice device2) {
+	long compareForIdentity(BlinkDevice device1, BlinkDevice device2) {
 		int mIdentityPointDev1 = device1.getIdentityPoint() ^ IDENTITY_POINTLINE_PROXY;
 		int mIdentityPointDev2 = device2.getIdentityPoint() ^ IDENTITY_POINTLINE_PROXY;
 		
@@ -298,7 +298,7 @@ public class DeviceAnalyzer {
 				
 			}
 			
-			return (int)(mTimestampDev1 - mTimestampDev2);
+			return (long)(mTimestampDev1 - mTimestampDev2);
 		}
 	}
